requirements.txt
anyio==3.6.2
fastapi==0.92.0
uvicorn==0.23.2
pydantic==1.10.5
pymongo==4.3.3
python-Levenshtein==0.20.9
dnspython==2.3.0
python-multipart==0.0.6
email-validator==2.0.0
python-jose==3.3.0
passlib==1.7.4
pytest==7.4.2
httpx==0.24.1
pip~=23.2.1
python-dotenv~=1.0.0